packaging/python/requirements/client.txt
Babel==1.3
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.18
boto==2.32.1
celery==3.1.13
certifi==14.05.14
coverage==3.7.1
futures==2.2.0
iso8601==0.1.10
kombu==3.0.21
mock==1.0.1
netaddr==0.7.12
nose==1.3.3
oslo.config==1.11.0
oslo.i18n==1.5.0
oslo.serialization==1.4.0
oslo.utils==1.4.0
pbr==0.11.0
prettytable==0.7.2
python-heatclient==0.2.12
python-keystoneclient==1.4.0
python-novaclient==2.24.1
python-swiftclient==2.3.1
pytz==2014.7
requests==2.3.0
simplejson==3.6.5
six==1.9.0
stevedore==1.4.0
tornado==4.0
wsgiref==0.1.2